fix: widen token columns to bigint, validate calendar dates and empty slugs, make report writes atomic
Fix round 1/5 for Task 6, addressing a review that found validation
diverging from what the storage layer actually accepts:

- tool_days and collective_days token columns (tokens_in, tokens_out,
  cache_read, cache_write) widen from integer to bigint (mode: 'number').
  int4 caps at 2,147,483,647; the collective rollup behind the homepage
  hero counter blows past that with only a few hundred developers.
  Migration 0000 regenerated in place (pre-launch, single migration).
- reportSchema's day field now also validates calendar validity, not
  just format, so "2026-13-99" is rejected by zod instead of reaching
  Postgres and surfacing as an unhandled 500.
- tool/model fields reject strings that pass .min(1) but slug down to
  an empty string (symbol-only or CJK-only input), which previously
  would have stored a blank aggregation bucket.
- The route's insert loop is now wrapped in a transaction (extracted
  into src/lib/ingest.ts's writeReport, so it's directly testable
  against an isolated in-memory pglite fixture) so a batch with one
  bad row commits nothing instead of leaving earlier rows in place.

MAX_TOKENS_PER_DAY stays at 10_000_000_000 — it's now a real bound
the DB honors rather than one Postgres silently overrode.

+++ drizzle/0000_overjoyed_ogun.sql
@@ -1,9 +1,9 @@
 CREATE TABLE "collective_days" (
 	"day" date PRIMARY KEY NOT NULL,
-	"tokens_in" integer DEFAULT 0 NOT NULL,
-	"tokens_out" integer DEFAULT 0 NOT NULL,
-	"cache_read" integer DEFAULT 0 NOT NULL,
-	"cache_write" integer DEFAULT 0 NOT NULL,
+	"tokens_in" bigint DEFAULT 0 NOT NULL,
+	"tokens_out" bigint DEFAULT 0 NOT NULL,
+	"cache_read" bigint DEFAULT 0 NOT NULL,
+	"cache_write" bigint DEFAULT 0 NOT NULL,
 	"cost_usd" numeric(14, 4) DEFAULT '0' NOT NULL
 );
 --> statement-breakpoint
@@ -32,10 +32,10 @@ CREATE TABLE "tool_days" (
 	"model" text NOT NULL,
 	"day" date NOT NULL,
 	"sessions" integer DEFAULT 0 NOT NULL,
-	"tokens_in" integer DEFAULT 0 NOT NULL,
-	"tokens_out" integer DEFAULT 0 NOT NULL,
-	"cache_read" integer DEFAULT 0 NOT NULL,
-	"cache_write" integer DEFAULT 0 NOT NULL,
+	"tokens_in" bigint DEFAULT 0 NOT NULL,
+	"tokens_out" bigint DEFAULT 0 NOT NULL,
+	"cache_read" bigint DEFAULT 0 NOT NULL,
+	"cache_write" bigint DEFAULT 0 NOT NULL,
 	"cost_usd" numeric(12, 4) DEFAULT '0' NOT NULL,
 	"source" text NOT NULL,
 	"verified" boolean DEFAULT false NOT NULL,
